+++ package.json
@@ -7,24 +7,22 @@
     "cypress:verify": "cypress verify",
     "start": "ibazel run src:devserver",
     "start-no-animations": "bazel run src:devserver-e2e",
-    "build": "ng build",
     "build:chrome": "bazel build projects/shell-chrome/src:prodapp",
     "build:chrome:watch": "ibazel build projects/shell-chrome/src:prodapp",
-    "release": "node projects/shell-chrome/set-version.js && cross-env LATEST_SHA=`git rev-parse --short HEAD` ng build shell-chrome --configuration production",
+    "release": "node projects/shell-chrome/set-version.js && cross-env LATEST_SHA=`git rev-parse --short HEAD` build:chrome",
     "test": "ng test",
     "lint": "tslint -p tsconfig.json -c tslint.json",
     "lint:fix": "tslint -p tsconfig.json -c tslint.json --fix",
-    "e2e": "npm run cy:run",
-    "cy:verify": "$(npm bin)/cypress verify",
-    "cy:version": "$(npm bin)/cypress version",
+    "e2e": "yarn cy:run",
+    "cy:verify": "$(yarn bin)/cypress verify",
+    "cy:version": "$(yarn bin)/cypress version",
     "cy:run": "$(yarn bin)/cypress run",
     "cy:open": "$(yarn bin)/cypress open",
     "cy:ci": "start-server-and-test start-no-animations http-get://localhost:4200 cy:run",
     "cy:watch": "start-server-and-test start-no-animations http-get://localhost:4200 cy:open",
     "test:ci": "bazel test projects/shared-utils/... projects/protocol/... projects/ng-devtools/... projects/ng-devtools-backend/...",
     "prettier": "prettier --write \"{,!(node_modules|dist|build|coverage)/**/}*.{js,jsx,ts,tsx,json}\"",
-    "prettier:fix": "pretty-quick --staged",
-    "bazel:build": "yarn bazel:build:protocol && yarn bazel:build:ng-devtools-backend && yarn bazel:build:shared-utils && yarn bazel:build:ng-devtools"
+    "prettier:fix": "pretty-quick --staged"
   },
   "husky": {
     "hooks": {
