+++ projects/ng-devtools-backend/index.ts
@@ -1,0 +1,1 @@
+export * from './src/public-api';
