+++ projects/ng-devtools-backend/src/lib/component-inspector/BUILD.bazel
@@ -1,0 +1,53 @@
+load("//tools:angular_ts_library.bzl", "ng_ts_library")
+load("@npm//@bazel/typescript:index.bzl", "ts_library")
+load("@npm//@bazel/concatjs:index.bzl", "karma_web_test_suite")
+
+package(default_visibility = ["//visibility:public"])
+
+ng_ts_library(
+    name = "component-inspector",
+    srcs = ["component-inspector.ts"],
+    deps = [
+        "//projects/protocol",
+        "//projects/ng-devtools-backend/src/lib:highlighter",
+        "//projects/ng-devtools-backend/src/lib:component-tree",
+        "//projects/ng-devtools-backend/src/lib:interfaces",
+        "//projects/ng-devtools-backend/src/lib/hooks",
+        "@npm//rxjs",
+        "@npm//@angular/core"
+    ],
+    tsconfig = "//projects/ng-devtools-backend:tsconfig-lib"
+)
+
+karma_web_test_suite(
+    name = "test",
+    browsers = [
+        "@io_bazel_rules_webtesting//browsers:chromium-local",
+    ],
+    tags = ["native"],
+    deps = [
+        ":test-lib",
+        "@npm//:node_modules/tslib/tslib.js",
+        "//src:rxjs_umd_modules",
+    ],
+)
+
+ts_library(
+    name = "test-lib",
+    testonly = True,
+    srcs = [
+        "component-inspector.spec.ts",
+    ],
+    deps = [
+        ":component-inspector",
+        "//projects/protocol",
+        "//projects/ng-devtools-backend/src/lib:component-tree",
+        "//projects/ng-devtools-backend/src/lib/hooks",
+        "//projects/ng-devtools-backend/src/lib:highlighter",
+        "@npm//@angular/core",
+        "@npm//@types",
+        "@npm//@types/jasmine",
+        "//projects/ng-devtools-backend/src/lib:interfaces",
+        "//projects/ng-devtools-backend/src/lib:utils",
+    ],
+)
